ref(media): upgrade media router to typescript
