+++ src/routes/v2/authenticatedSites/index.js
@@ -10,10 +10,7 @@ const {
 } = require("@routes/v2/authenticatedSites/collections")
 const { ContactUsRouter } = require("@routes/v2/authenticatedSites/contactUs")
 const { HomepageRouter } = require("@routes/v2/authenticatedSites/homepage")
-const {
-  MediaCategoriesRouter,
-} = require("@routes/v2/authenticatedSites/mediaCategories")
-const { MediaFilesRouter } = require("@routes/v2/authenticatedSites/mediaFiles")
+const { MediaRouter } = require("@routes/v2/authenticatedSites/media")
 const { NavigationRouter } = require("@routes/v2/authenticatedSites/navigation")
 const {
   RepoManagementRouter,
@@ -182,11 +179,9 @@ const getAuthenticatedSitesSubrouter = ({
   const resourceDirectoryV2Router = new ResourceCategoriesRouter({
     resourceDirectoryService,
   })
-  const mediaFilesV2Router = new MediaFilesRouter({
-    mediaFileService,
-  })
-  const mediaDirectoryV2Router = new MediaCategoriesRouter({
+  const mediaV2Router = new MediaRouter({
     mediaDirectoryService,
+    mediaFileService,
   })
   const resourceRoomV2Router = new ResourceRoomRouter({
     resourceRoomDirectoryService,
@@ -231,11 +226,7 @@ const getAuthenticatedSitesSubrouter = ({
     "/resourceRoom/:resourceRoomName/resources",
     resourceDirectoryV2Router.getRouter()
   )
-  authenticatedSitesSubrouter.use(
-    "/media/:directoryName/pages",
-    mediaFilesV2Router.getRouter()
-  )
-  authenticatedSitesSubrouter.use("/media", mediaDirectoryV2Router.getRouter())
+  authenticatedSitesSubrouter.use("/media", mediaV2Router.getRouter())
   authenticatedSitesSubrouter.use("/navigation", navigationV2Router.getRouter())
   authenticatedSitesSubrouter.use(
     "/resourceRoom",
